fix: syscall 5 problem is now kinda solved

+++ src/lib-header/terminal.h
@@ -3,10 +3,10 @@
 
 #include "framebuffer.h"
 
-extern const size_t VGA_WIDTH;
-extern const size_t VGA_HEIGHT;
-extern size_t row;
-extern size_t column;
+extern const uint8_t VGA_WIDTH;
+extern const uint8_t VGA_HEIGHT;
+extern uint8_t row;
+extern uint8_t column;
 extern uint8_t background_color;
 
 void init_terminal(void);
